# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateScheduledTransferBody": {
        "additionalProperties": false,
        "properties": {
          "amountPoisha": {
            "exclusiveMinimum": 0.0,
            "title": "Amountpoisha",
            "type": "integer"
          },
          "executeAt": {
            "format": "date-time",
            "title": "Executeat",
            "type": "string"
          },
          "note": {
            "anyOf": [
              {
                "maxLength": 140,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "pin": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pin"
          },
          "recipientPhone": {
            "title": "Recipientphone",
            "type": "string"
          }
        },
        "required": [
          "recipientPhone",
          "amountPoisha",
          "executeAt"
        ],
        "title": "CreateScheduledTransferBody",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RecipientBody": {
        "additionalProperties": false,
        "properties": {
          "amountPoisha": {
            "title": "Amountpoisha",
            "type": "integer"
          },
          "phone": {
            "title": "Phone",
            "type": "string"
          }
        },
        "required": [
          "phone",
          "amountPoisha"
        ],
        "title": "RecipientBody",
        "type": "object"
      },
...
      "ScheduledRecipient": {
        "properties": {
          "maskedPhone": {
            "title": "Maskedphone",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "maskedPhone"
        ],
        "title": "ScheduledRecipient",
        "type": "object"
      },
      "ScheduledTransferList": {
        "properties": {
          "scheduledTransfers": {
            "items": {
              "$ref": "#/components/schemas/ScheduledTransferResource"
            },
            "title": "Scheduledtransfers",
            "type": "array"
          }
        },
        "required": [
          "scheduledTransfers"
        ],
        "title": "ScheduledTransferList",
        "type": "object"
      },
      "ScheduledTransferResource": {
        "properties": {
          "amountPoisha": {
            "title": "Amountpoisha",
            "type": "integer"
          },
          "authorizedAt": {
            "format": "date-time",
            "title": "Authorizedat",
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "executeAt": {
            "format": "date-time",
            "title": "Executeat",
            "type": "string"
          },
          "failureCode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failurecode"
          },
          "failureMessage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failuremessage"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "recipient": {
            "$ref": "#/components/schemas/ScheduledRecipient"
          },
          "reference": {
            "title": "Reference",
            "type": "string"
          },
          "resolvedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolvedat"
          },
          "scheduledTransferId": {
            "title": "Scheduledtransferid",
            "type": "string"
          },
          "status": {
            "enum": [
              "SCHEDULED",
              "EXECUTED",
              "FAILED",
              "CANCELLED"
            ],
            "title": "Status",
            "type": "string"
          },
          "transferReference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transferreference"
          }
        },
        "required": [
          "scheduledTransferId",
          "reference",
          "status",
          "amountPoisha",
          "note",
          "executeAt",
          "recipient",
          "transferReference",
          "failureCode",
          "failureMessage",
          "authorizedAt",
          "resolvedAt",
          "createdAt"
        ],
        "title": "ScheduledTransferResource",
        "type": "object"
      },
...
      "TransferBody": {
        "additionalProperties": false,
        "description": "Accepts the one-to-one shorthand and the group form.\n\nBoth snake_case and camelCase are accepted on input. This is not indecision --\nthe frontend and the load tests were written against different conventions, and\na field-name mismatch discovered at demo time is a worse outcome than an alias.",
        "properties": {
          "amountPoisha": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amountpoisha"
          },
          "note": {
            "anyOf": [
              {
                "maxLength": 140,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "pin": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pin"
          },
          "recipientPhone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recipientphone"
          },
          "recipients": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/RecipientBody"
                },
                "maxItems": 20,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recipients"
          }
        },
        "title": "TransferBody",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Money Movement",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/scheduled-transfers": {
      "get": {
        "operationId": "list_scheduled_transfers_api_v1_scheduled_transfers_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduledTransferList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Scheduled Transfers",
        "tags": [
          "scheduled-transfers"
        ]
      },
      "post": {
        "operationId": "create_scheduled_transfer_api_v1_scheduled_transfers_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateScheduledTransferBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduledTransferResource"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Scheduled Transfer",
        "tags": [
          "scheduled-transfers"
        ]
      }
    },
    "/api/v1/scheduled-transfers/{schedule_id}": {
      "get": {
        "operationId": "get_scheduled_transfer_api_v1_scheduled_transfers__schedule_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "schedule_id",
            "required": true,
            "schema": {
              "title": "Schedule Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduledTransferResource"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Scheduled Transfer",
        "tags": [
          "scheduled-transfers"
        ]
      }
    },
    "/api/v1/scheduled-transfers/{schedule_id}/cancel": {
      "post": {
        "operationId": "cancel_scheduled_transfer_api_v1_scheduled_transfers__schedule_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "schedule_id",
            "required": true,
            "schema": {
              "title": "Schedule Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduledTransferResource"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Scheduled Transfer",
        "tags": [
          "scheduled-transfers"
        ]
      }
    },
...
    "/api/v1/transfers": {
      "get": {
        "description": "This Account's movements, read from the Ledger rather than from transfers.\n\nReading the journal is what makes a Group Transfer show up correctly for each\nrecipient: they see their own leg, not the sender's total.",
        "operationId": "list_transfers_api_v1_transfers_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "direction",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "sent",
                    "received"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Direction"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Transfers",
        "tags": [
          "transfers"
        ]
      },
      "post": {
        "operationId": "create_transfer_api_v1_transfers_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransferBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Transfer",
        "tags": [
          "transfers"
        ]
      }
    },
    "/api/v1/transfers/{reference}": {
      "get": {
        "operationId": "get_transfer_api_v1_transfers__reference__get",
        "parameters": [
          {
            "in": "path",
            "name": "reference",
            "required": true,
            "schema": {
              "title": "Reference",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Transfer",
        "tags": [
          "transfers"
        ]
      }
    },
...
  }
}